ProtoGlod: Publishing the repo!
The very first commit omg!

I created a README file and put down my thoughts about the initial idea for the project; drafted a manifesto/ artist statement of sorts.

I still need to expand on the specifics of the book's mechanics, I feel like. Maybe a more dedicated design document is called for at some point in the future.

And the structure of the repo at this point is still pretty bare bones. I've added an ideation folder with two moodboards and a subfolder for concept art.

the 'narrative concept' moodboard was made with Midjourney AI and is a very rough idea of the style. I'll use it as a jumping-off point to work actual concept art that I cam work off as I progress through the different aspects of the game.