ProtoGlod: Project structure set, first journal entry, now what?
Added the UE5 project folder, and the process folder with all the MDM tracking files.

Wrote a journal entry trying to trace back to the beginning of the idea for this project, and projecting my excitement for the future.

I gotta decide where to start now haha! Do I conceptualize some more? Make some concept art? Or do I jump straight into battling with the tech to get the core mechanics built?